requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
minimalmodbus==1.0.2
numpy==1.18.1
pandas==1.0.1
pyserial==3.4
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
